Fix codacy's recommendations

+++ src/main/java/org/chocosolver/solver/constraints/nary/PropKnapsack.java
@@ -105,8 +105,8 @@ public class PropKnapsack extends Propagator<IntVar> {
         double camax = capacity.getUB();
         double pomin = 0;
         for (int i = 0; i < n; i++) {
-            camax -= weigth[i] * vars[i].getLB();
-            pomin += energy[i] * vars[i].getLB();
+            camax -= weigth[i] * (long)vars[i].getLB();
+            pomin += energy[i] * (long)vars[i].getLB();
         }
         if (camax < 0 || pomin > power.getUB()) {
             return ESat.FALSE;
